backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
SQLAlchemy==2.0.30
python-multipart==0.0.9
python-dotenv==1.0.1
PyYAML==6.0.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.0
pydantic-settings==2.5.2
markdown2==2.4.12
email-validator==2.2.0
bcrypt==4.0.1
google-generativeai==0.8.3
